Feature/16-Frontend-Post (#19)
Add the ability to handle a POST request from the frontend.

---
* Send CSRF, Fix 500 Error, Update gitignore

* Get title from Google Books API

* Include time when creating Listing

* Bulk Update - Multiple images can now be uploaded and saved in the DB.

* Update Travis JDK to 11

* Update tests to work with new code

- Commented out two tests, noted as a minor issue
- Downgraded hashcode function, noted as minor issue

* Fix SonarQube Code Smells - Optimized imports and removed comment blocks

* Small fix to get tests working again

* Moved RestTemplate to class scope to fix last two tests

* Add newListing test and move model/entity conversion out of controller

* Travis back to JDK 8

* Travis to OpenJDK 11

* Fix NullPointerException

* Trying Travis OpenJDK8

* Fix broken Rest Template.

+++ src/main/java/com/gmu/bookshare/persistence/ListingRepository.java
@@ -9,5 +9,5 @@ public interface ListingRepository extends JpaRepository<ListingEntity, Long> {
 
     List<ListingEntity> findByTitle(String s);
 
-    List<ListingEntity> findByIsbn(int isbn);
+    List<ListingEntity> findByIsbn(long isbn);
 }
